BugID: 18991432: Fix up AOS WhiteScan bugs
Reviewed-on: http://yuncode.alibaba-inc.com/195788
Tested-by: 李诚 lc122798 <lc122798@alibaba-inc.com>

+++ security/das/src/board/mk3080/das_sha1.c
@@ -33,7 +33,7 @@ void das_sha1_start(das_sha1_context_t *ctx)
 
     ctx->context = aos_malloc(ctx->size);
     if (!ctx->context) {
-        printf("malloc %d fail\n", ctx->size);
+        printf("malloc %d fail\n", (int)ctx->size);
         DAS_SHA1_ASSERT(0);
     }
